+++ src/installNpmPkgs.js
@@ -2,31 +2,31 @@
  * install npm packages
  * @author JA+
  */
-const childProcess = require('child_process')
-const { npmPkgs } = require('./config')
+const childProcess = require('child_process');
+const { npmPkgs } = require('./config');
 
 module.exports = function ({ type }) {
-  let npmCmd = 'npm i -D --legacy-peer-deps'
+  let npmCmd = 'npm i -D --legacy-peer-deps';
 
-  const pkgTypes = ['eslint']
+  const pkgTypes = ['eslint'];
   if (type !== 'svelte3') {
     // svelte not support prettier
-    pkgTypes.push('prettier')
+    pkgTypes.push('prettier');
   }
-  pkgTypes.push(type)
+  pkgTypes.push(type);
 
   for (const key of pkgTypes) {
-    const pkgs = npmPkgs[key]
+    const pkgs = npmPkgs[key];
     for (const pkg in pkgs) {
-      const version = pkgs[pkg]
-      npmCmd += ' ' + pkg + '@' + version
+      const version = pkgs[pkg];
+      npmCmd += ' ' + pkg + '@' + version;
     }
   }
-  console.log(npmCmd)
+  console.log(npmCmd);
   try {
-    childProcess.execSync(npmCmd)
-    console.log('✔ Npm packages installation succeed')
+    childProcess.execSync(npmCmd);
+    console.log('✔ Npm packages installation succeed');
   } catch (err) {
-    if (err) return console.error('✘ Npm install failed', err)
+    if (err) return console.error('✘ Npm install failed', err);
   }
-}
+};