+++ src/main.js
@@ -3,16 +3,18 @@
  * Entry
  * @author JA+
  */
-const childProcess = require('child_process')
-const inquirer = require('inquirer')
-const { promptChoices, minCodeVersion } = require('./config.js')
-const prompt = inquirer.createPromptModule()
+const childProcess = require('child_process');
+const pkg = require('../package.json');
+const chalk = require('chalk');
+const inquirer = require('inquirer');
+const { promptChoices, minCodeVersion } = require('./config.js');
+const prompt = inquirer.createPromptModule();
 
-let vscodeVersion = childProcess.execSync('code --version').toString()
-vscodeVersion = vscodeVersion.split('\n')[0]
+let vscodeVersion = childProcess.execSync('code --version').toString();
+vscodeVersion = vscodeVersion.split('\n')[0];
 
-let promptParams = []
-
+let promptParams = [];
+console.log('version:', chalk.green(pkg.version));
 // Check vscode version
 if (vscodeVersion < minCodeVersion) {
   promptParams.push({
@@ -20,7 +22,7 @@ if (vscodeVersion < minCodeVersion) {
     name: 'invalidVersion',
     message: `Your vscode version(${vscodeVersion}) < ${minCodeVersion}, It can cause some errors, Still continue?`,
     default: false,
-  })
+  });
 }
 
 promptParams.push({
@@ -30,32 +32,32 @@ promptParams.push({
   message: 'Select env:',
   choices: promptChoices,
   when(answer) {
-    return answer.invalidVersion === undefined || answer.invalidVersion === true
+    return answer.invalidVersion === undefined || answer.invalidVersion === true;
   },
-})
+});
 
 prompt(promptParams).then(async answer => {
-  if (answer.invalidVersion === false) return
+  if (answer.invalidVersion === false) return;
 
   try {
-    console.log('\n► Start Installing vscode extension.')
-    require('./installExt.js')(answer)
+    console.log('\n► Start Installing vscode extension.');
+    require('./installExt.js')(answer);
 
     // update settings.json
-    console.log('\n► Vscode settings.json add configuration...')
-    require('./updateSettings.js')(answer)
+    console.log('\n► Vscode settings.json add configuration...');
+    require('./updateSettings.js')(answer);
 
-    console.log('\n► Copying config files...')
-    await require('./copyConfigFile.js')(answer)
+    console.log('\n► Copying config files...');
+    await require('./copyConfigFile.js')(answer);
 
-    console.log('\n► Installing npm packages...')
-    require('./installNpmPkgs.js')(answer)
+    console.log('\n► Installing npm packages...');
+    require('./installNpmPkgs.js')(answer);
 
-    console.log('\n✔ All task done. Please restart vscode( / Restart eslint plugin / Reload require). Make effective eslint.')
+    console.log('\n✔ All task done. Please restart vscode( / Restart eslint plugin / Reload require). Make effective eslint.');
     console.log(
       '\n? Did not come into effect? Make sure the folder that vscode opened, which has the config file and node_modules in root directory.',
-    )
+    );
   } catch (err) {
-    console.error(err)
+    console.error(err);
   }
-})
+});