+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 1.4.5
+* add: version log
+* add: `chalk` highlight log
 ## 1.4.4
 * support vue2.7 setup macros `defineProps` `defineEmits` `defineExpose` `withDefaults`
 * upgrade dependencies `eslint-plugin-vue` `vue-eslint-parser`